Separate logic in threads
